refactor(st-09055): modularize schema inspector runtime
